Rename: 폴더 구조 및 파일 위치 정리
